+++ src/main/java/gift/model/Product.java
@@ -1,0 +1,51 @@
+package gift.model;
+
+public class Product {
+
+  private Long id;
+  private String name;
+  private int price;
+  private String imageUrl;
+
+  public Product() {
+  }
+
+  public Product(Long id, String name, int price, String imageUrl) {
+    this.id = id;
+    this.name = name;
+    this.price = price;
+    this.imageUrl = imageUrl;
+  }
+
+  public Long getId() {
+    return id;
+  }
+
+  public void setId(Long id) {
+    this.id = id;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public void setName(String name) {
+    this.name = name;
+  }
+
+  public int getPrice() {
+    return price;
+  }
+
+  public void setPrice(int price) {
+    this.price = price;
+  }
+
+  public String getImageUrl() {
+    return imageUrl;
+  }
+
+  public void setImageUrl(String imageUrl) {
+    this.imageUrl = imageUrl;
+  }
+}
